Exit if no output file name is given

+++ tools/elf2e32/elf2e32.cpp
@@ -462,6 +462,12 @@ int main(int argc, char *argv[]) {
 		return 0;
 	}
 
+	if (!output) {
+		printf("no output file specified\n");
+		free(dso);
+		return 1;
+	}
+
 	int fd = open(elfinput, O_RDONLY);
 	unlink(output);
 	FILE* out = fopen(output, "w+b");
